Пришлось объеденить все в 1 файл, Задание 1 Вариант 3
