--- kuznetsov.alexey/T2/local.h
@@ -1,18 +1,0 @@
-#ifndef LOCAL_H
-#define LOCAL_H
-
-
-
-#include <iostream>
-#include <sstream>
-#include <string>
-#include <cassert>
-#include <iterator>
-#include <vector>
-#include <iomanip>
-#include <complex>
-#include <algorithm>
-#include <cmath>
-#include <exception>
-
-#endif#
